refactor: move include guard logic to lib.rs and clean up main.rs
- Extract include guard generation, language, and line ending enums to src/lib.rs
- Update main.rs to use the new guardgen module
- Remove duplicate code and tests from main.rs
- Add src/lib.rs with documentation and tests
